# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Gratus API",
        "version": "0.7"
    },
    "paths": {
...
        "/api/mobile/forgotPass": {
            "post": {
                "tags": [
                    "Mobile"
                ],
                "summary": "Forgot Password",
                "description": "",
                "operationId": "forgotPass",
                "requestBody": {
                    "description": "User credentials",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [],
                                "properties": {
                                    "email": {
                                        "type": "email",
                                        "example": "email@gratus.user"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "No body in answer"
                    },
                    "543": {
                        "description": "Email is incorrect."
                    },
                    "544": {
                        "description": "Email should contain @ and ."
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
...
        "/api/mobile/forgotPassConfirm": {
            "post": {
                "tags": [
                    "Mobile"
                ],
                "summary": "Password recovery confirm",
                "description": "",
                "operationId": "forgotPassConfirm",
                "requestBody": {
                    "description": "User credentials",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [],
                                "properties": {
                                    "email": {
                                        "type": "email",
                                        "example": "email@gratus.user"
                                    },
                                    "password": {
                                        "type": "password",
                                        "example": "password"
                                    },
                                    "repeat_password": {
                                        "type": "password",
                                        "example": "password"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Password successfully changed"
                    },
                    "545": {
                        "description": "Password should contain at least 6 characters"
                    },
                    "546": {
                        "description": "Passwords don't match"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
...
    },
    "security": [
        []
    ]
}